Stealing win_ec2config module from ansible-modules-extra PR
See https://github.com/ansible/ansible-modules-extras/pull/442

The module is not done and is tagged for revision, but does what we want
now, so I manually included it.  If it gets included (probably won't get
back ported, so we won't see it until we get to 2.0), we will just
remove this.