Change iam and iam_policy_list in to the obliterating modules
This diff will look ugly.  It's not meant to be a fork.  The original
IAM and IAM policy modules give no option to A) list IAM objects or B)
purge unlisted ones.  Thus the following modules (which are rename in
the next commit), iam_obliterator and iam_policy_obliterator.

In summary when provided with a spare list and a path
(http://docs.aws.amazon.com/IAM/latest/UserGuide/reference_identifiers.html#identifiers-friendly-names)
the obliterator module will go and destroy anything A) Having the given
path AND NOT B) being in the spare list (by name).  Use with care.