Rename iam|_policy_list.py to *_obliterator.py
Moving new modules to their own names.  Recall, they are not forks of
the prevous names, only wrought from them.

Ignore the common/deploy/ madness.  This was due to me thinking we were
going to symlink library instead of making it a real boy^H^H^Hdirectory.
Going to leave these, because I will be rebasing on top of master were
iam_policy (in the proper library/ location) is written to, so I don't
want to mess with that.