Merging upstream fix for ec2_elb_lb
Applying https://github.com/ansible/ansible-modules-core/pull/3595 to
our fork of ec2_elb_lb.  This is merged upstream.  In short, the
ec2_elb_lb module would fail out, but the ELBs would be created.
However, I noticed that tags were not being applied until the second run
of the playbook, so the bug was not a benign as I originally though.

We are already forking said module, so I applied the fix manually.  When
we move to the mainstream version of the module, this fix will be
included.